requirements.txt
attrs==23.1.0
aws-cdk-lib==2.81.0
aws-cdk.asset-awscli-v1==2.2.186
aws-cdk.asset-kubectl-v20==2.1.1
aws-cdk.asset-node-proxy-agent-v5==2.0.158
boto3==1.26.150
botocore==1.29.150
cattrs==22.2.0
constructs==10.2.46
exceptiongroup==1.1.1
importlib-metadata==6.6.0
importlib-resources==5.12.0
jmespath==1.0.1
jsii==1.83.0
publication==0.0.3
python-dateutil==2.8.2
python-decouple==3.8
s3transfer==0.6.1
six==1.16.0
typeguard==2.13.3
typing_extensions==4.6.3
urllib3==1.26.16
zipp==3.15.0
